Adds zero downtime deployment

+++ scripts/build.sh
@@ -2,15 +2,31 @@
 
 cd "$(dirname "$0")/.."
 
-rm -rf ./build/
+rm -rf ./build-new/
 
 webpack
 
-cp -r src/api/ build/
-cp -r src/json/ build/
-cp -r src/favicon.ico build/
-cp -r src/opengraph.png build/
+cp -r src/api/ build-new/
+cp -r src/json/ build-new/
+cp -r src/favicon.ico build-new/
+cp -r src/opengraph.png build-new/
 
 php ./composer.phar install
-php ./scripts/generate-results-json.php
 
+du -sh build-new/ build/
+
+while true; do
+    read -p "Deploy new build to production? " yn
+    case $yn in
+        [Yy]* )
+          rm -rf ./build/
+          mv ./build-new/ build/
+        break;;
+
+        [Nn]* )
+          echo "Aborted."
+        exit;;
+
+        * ) echo "Please answer yes or no.";;
+    esac
+done
